backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
flask-mail==0.9.1
python-dotenv==1.0.0
bcrypt==4.0.1
sqlalchemy==2.0.20
